Change: Rename selection type module to snake-case
Only modules containing react components should use camelCase.
