+++ src/web/entities/EntitiesContainer.tsx
@@ -43,7 +43,7 @@ import {loadUserSettingDefaults} from 'web/store/usersettings/defaults/actions';
 import {getUserSettingsDefaults} from 'web/store/usersettings/defaults/selectors';
 import compose from 'web/utils/compose';
 import {generateFilename} from 'web/utils/Render';
-import SelectionType, {type SelectionTypeType} from 'web/utils/SelectionType';
+import SelectionType, {type SelectionTypeType} from 'web/utils/selection-type';
 import SortDirection, {type SortDirectionType} from 'web/utils/sort-direction';
 import {withRouter} from 'web/utils/withRouter';
 import withTranslation from 'web/utils/withTranslation';
