+++ src/web/pages/notes/__tests__/NoteTableHeader.test.tsx
@@ -6,7 +6,7 @@
 import {describe, expect, test, testing} from '@gsa/testing';
 import {rendererWithTable, screen, fireEvent} from 'web/testing';
 import NoteTableHeader from 'web/pages/notes/NoteTableHeader';
-import SelectionType from 'web/utils/SelectionType';
+import SelectionType from 'web/utils/selection-type';
 
 describe('NoteTableHeader tests', () => {
   test('should render the table header', () => {
