+++ src/web/utils/selection-type.ts
